reference-app/trial/requirements.txt
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
jaeger-client==4.3.0
opentelemetry-api==0.12b0
opentelemetry-sdk==0.12b0
opentelemetry-instrumentation==0.12b0
opentelemetry-instrumentation-flask==0.12b0
opentelemetry-instrumentation-requests==0.12b0
opentelemetry-exporter-jaeger==0.12b0
prometheus_client==0.12.0
prometheus-flask-exporter==0.18.3
Flask-Cors==3.0.10
Jinja2==3.0.3
itsdangerous==2.0.1
werkzeug==2.0.3
Flask-OpenTracing==1.1.0
